move majorana_operator to _fermion_operator.py and expose it
